refactor: move GraphQL query tests to specific files
